+++ resources/views/bootstrap-4/includes/pagination.blade.php
@@ -1,4 +1,4 @@
-@if ($showPagination)
+@if ($paginationEnabled && $showPerPage)
     <div class="row">
         <div class="col-12 col-md-6">
             {{ $rows->links() }}