+++ resources/views/bootstrap-5/includes/per-page.blade.php
@@ -1,11 +1,13 @@
-@if ($showPerPage)
-    <select
-        wire:model="perPage"
-        id="perPage"
-        class="form-select"
-    >
-        <option value="10">10</option>
-        <option value="25">25</option>
-        <option value="50">50</option>
-    </select>
+@if ($paginationEnabled && $showPerPage)
+    <div class="ms-0 ms-md-3">
+        <select
+            wire:model="perPage"
+            id="perPage"
+            class="form-select"
+        >
+            @foreach ($perPageAccepted as $item)
+                <option value="{{ $item }}">{{ $item }}</option>
+            @endforeach
+        </select>
+    </div>
 @endif