+++ resources/views/tailwind/includes/pagination.blade.php
@@ -1,4 +1,4 @@
-@if ($showPagination)
+@if ($paginationEnabled && $showPerPage)
     <div class="p-6 md:p-0">
         {{ $rows->links() }}
     </div>