+++ src/Traits/WithBulkActions.php
@@ -2,11 +2,15 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
+use Illuminate\Database\Eloquent\Builder;
+
 /**
  * Trait WithBulkActions.
  */
 trait WithBulkActions
 {
+    public string $primaryKey = 'id';
+    public bool $showFilters = true;
     public bool $selectPage = false;
     public bool $selectAll = false;
     public $selected = [];
@@ -39,7 +43,7 @@ trait WithBulkActions
 
     public function selectPageRows(): void
     {
-        $this->selected = $this->rows->pluck('id')->map(fn ($id) => (string) $id);
+        $this->selected = $this->rows->pluck($this->primaryKey)->map(fn ($key) => (string) $key);
     }
 
     public function selectAll(): void
@@ -54,9 +58,24 @@ trait WithBulkActions
         $this->selected = [];
     }
 
-    public function getSelectedRowsQueryProperty()
+    public function selectedRowsQuery(): Builder
     {
-        return (clone $this->rowsQuery)
+        return (clone $this->rowsQuery())
             ->unless($this->selectAll, fn ($query) => $query->whereKey($this->selected));
+    }
+
+    public function getSelectedRowsQueryProperty(): Builder
+    {
+        return $this->selectedRowsQuery();
+    }
+
+    public function selectedKeys(): array
+    {
+        return $this->selectedRowsQuery()->pluck($this->primaryKey)->toArray();
+    }
+
+    public function getSelectedKeysProperty(): array
+    {
+        return $this->selectedKeys();
     }
 }