+++ src/Traits/WithPerPagePagination.php
@@ -7,8 +7,11 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
  */
 trait WithPerPagePagination
 {
+    public bool $paginationEnabled = true;
+    public bool $showPerPage = true;
+    public bool $showPagination = true;
     public int $perPage = 10;
-    protected array $perPageAccepted = [10, 25, 50];
+    public array $perPageAccepted = [10, 25, 50];
 
     public function mountWithPerPagePagination(): void
     {
@@ -19,6 +22,9 @@ trait WithPerPagePagination
         }
     }
 
+    /**
+     * @param $value
+     */
     public function updatedPerPage($value): void
     {
         if (in_array(session()->get($this->tableName.'-perPage', $this->perPage), $this->perPageAccepted, true)) {
@@ -30,6 +36,10 @@ trait WithPerPagePagination
         $this->resetPage();
     }
 
+    /**
+     * @param $query
+     * @return mixed
+     */
     public function applyPagination($query)
     {
         return $query->paginate($this->perPage, ['*'], $this->pageName());