+++ src/Traits/WithSorting.php
@@ -9,8 +9,10 @@ use Illuminate\Database\Eloquent\Builder;
  */
 trait WithSorting
 {
+    public bool $showSorting = true;
     public array $sorts = [];
     public array $sortNames = [];
+    public array $sortDirectionNames = [];
 
     public function sortBy(string $field): ?string
     {