+++ tests/DataTableComponentTest.php
@@ -3,6 +3,7 @@
 namespace Rappasoft\LaravelLivewireTables\Tests;
 
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Collection;
 use Rappasoft\LaravelLivewireTables\DataTableComponent;
 use Rappasoft\LaravelLivewireTables\Tests\Http\Livewire\PetsTable;
 
@@ -38,6 +39,33 @@ class DataTableComponentTest extends TestCase
 
         $this->assertInstanceOf(LengthAwarePaginator::class, $rows);
         $this->assertEquals(5, $this->table->getRowsProperty()->total());
+    }
+
+    /** @test */
+    public function test_pagination_default(): void
+    {
+        $this->assertInstanceOf(LengthAwarePaginator::class, $this->table->rows);
+        $this->assertEquals(10, $this->table->perPage);
+        $this->assertTrue($this->table->paginationEnabled);
+        $this->assertTrue($this->table->showPerPage);
+    }
+
+    /** @test */
+    public function test_pagination(): void
+    {
+        $this->table->perPage = 2;
+        $this->assertEquals(1, $this->table->rows->currentPage());
+        $this->assertEquals(2, $this->table->rows->count());
+        $this->assertEquals(3, $this->table->rows->lastPage());
+    }
+
+    /** @test */
+    public function test_pagination_disabled(): void
+    {
+        $this->table->paginationEnabled = false;
+        $this->table->perPage = 2;
+        $this->assertInstanceOf(Collection::class, $this->table->rows);
+        $this->assertCount(5, $this->table->rows);
     }
 
     /** @test */